backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-docx
openai
python-dotenv
python-jose
pydantic-settings
openpyxl
pypdf
requests
beautifulsoup4
anthropic>=0.40.0
langchain
langchain-openai
langchain-anthropic
langgraph
langgraph-checkpoint
tiktoken
langsmith
langgraph-checkpoint-sqlite
langgraph-prebuilt
alembic
deepl
jinja2